Added tests to interp1d fix
