+++ kaska/interp_fix.py
@@ -28,57 +28,3 @@ def interp1d(newx, oldx, oldy):
             newy[i] = np.interp(newx, oldx[mask], y[mask])
     newy_T = newy.transpose(1,0).reshape((new_shape,) + oldy_shape)
     return newy_T
-
-def test1d():
-    newx = np.arange(200)
-    oldx = np.array(sorted(np.random.choice(np.arange(200), 100, replace=False)))
-    oldy = oldy = np.random.rand(100)
-    np_ret    = np.interp(newx, oldx, oldy)
-    numba_ret = interp1d(newx, oldx, oldy)
-    if not np.allclose(np_ret, numba_ret):
-        raise
-    print('Same result achieved')
-
-def test2d():
-    newx = np.arange(200)
-    oldx = np.array(sorted(np.random.choice(np.arange(200), 100, replace=False)))
-    oldy = oldy = np.random.rand(100,5)
-    numba_ret = interp1d(newx, oldx, oldy)
-    for i in range(5):
-        np_ret    = np.interp(newx, oldx, oldy[:,i])
-        if not np.allclose(np_ret, numba_ret[:,i]):
-            raise
-    print('Same result achieved')
-
-def test3d():
-    newx = np.arange(200)
-    oldx = np.array(sorted(np.random.choice(np.arange(200), 100, replace=False)))
-    oldy = oldy = np.random.rand(100, 5, 10)
-    numba_ret = interp1d(newx, oldx, oldy)
-    for i in range(5):
-        for j in range(10):
-            np_ret    = np.interp(newx, oldx, oldy[:,i, j])
-            if not np.allclose(np_ret, numba_ret[:,i, j]):
-                raise
-    print('Same result achieved')
-
-def testgap():
-    newx = np.arange(200)
-    oldx = np.array(sorted(np.random.choice(np.arange(200), 100, replace=False)))
-    oldy = oldy = np.random.rand(100, 5, 10)
-    for i in range(5):
-        for j in range(10):
-            bad_pix = np.random.choice(range(100), 100)
-            gap_oldy = oldy[:,i, j].copy()
-            gap_oldy[bad_pix] = np.nan
-            numba_ret = interp1d(newx, oldx, gap_oldy)
-            np_ret = np.interp(newx, oldx, gap_oldy)
-            diff    = numba_ret - np_ret
-            if not np.nansum(diff)<1e-10:
-                raise
-    print('Same result achieved and filled gaps')
-if __name__ == '__main__':
-    test1d()
-    test2d() 
-    test3d()
-    testgap()
