+++ kaska/tests/test_interp1d.py
@@ -1,0 +1,59 @@
+'''
+Test the smoothn functionality, including the examples from the Matlab original
+
+Author: Timothy Spain, t.spain@ucl.ac.uk
+Date: 5th July 2019
+
+'''
+
+import pytest
+import numpy as np
+
+import sys
+
+from .. import interp_fix
+
+
+
+def test1d():
+    newx = np.arange(200)
+    oldx = np.array(sorted(np.random.choice(np.arange(200), 100, replace=False)))
+    oldy = oldy = np.random.rand(100)
+    np_ret    = np.interp(newx, oldx, oldy)
+    numba_ret = interp_fix.interp1d(newx, oldx, oldy)
+    assert np.allclose(np_ret, numba_ret)
+
+def test2d():
+    newx = np.arange(200)
+    oldx = np.array(sorted(np.random.choice(np.arange(200), 100, replace=False)))
+    oldy = oldy = np.random.rand(100,5)
+    numba_ret = interp_fix.interp1d(newx, oldx, oldy)
+    for i in range(5):
+        np_ret    = np.interp(newx, oldx, oldy[:,i])
+        assert np.allclose(np_ret, numba_ret[:,i])
+
+def test3d():
+    newx = np.arange(200)
+    oldx = np.array(sorted(np.random.choice(np.arange(200), 100, replace=False)))
+    oldy = oldy = np.random.rand(100, 5, 10)
+    numba_ret = interp_fix.interp1d(newx, oldx, oldy)
+    for i in range(5):
+        for j in range(10):
+            np_ret    = np.interp(newx, oldx, oldy[:,i, j])
+            assert np.allclose(np_ret, numba_ret[:,i, j])
+
+# def testgap():
+#     newx = np.arange(200)
+#     oldx = np.array(sorted(np.random.choice(np.arange(200), 100, replace=False)))
+#     oldy = oldy = np.random.rand(100, 5, 10)
+#     for i in range(5):
+#         for j in range(10):
+#             bad_pix = np.random.choice(range(100), 100)
+#             gap_oldy = oldy[:,i, j].copy()
+#             gap_oldy[bad_pix] = np.nan
+#             numba_ret = interp_fix.interp1d(newx, oldx, gap_oldy)
+#             np_ret = np.interp(newx, oldx, gap_oldy)
+#             diff    = numba_ret - np_ret
+#             if not np.nansum(diff)<1e-10:
+#                 raise
+#     print('Same result achieved and filled gaps')
